@@ -7,9 +7,9 @@
   },
   "dependencies": {
     "lodash.throttle": "^4.1.1",
-    "next": "latest",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "next": "^13.2.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-intersection-observer": "^8.33.1"
   },
   "devDependencies": {
