@@ -1,5 +1,5 @@
 import { Seiyuu } from 'api/anilist/types'
-import Card from 'components/shared/Card'
+import Card from '../base-ui/Card'
 
 type Props = {
   seiyuu: Seiyuu
@@ -16,7 +16,7 @@ export default function SeiyuuCard({ seiyuu }: Props) {
         <span>
           <div className="text-textSecondary">Voiced by</div>
           <a
-            className="text-textPrimary hover:text-primary text-lg"
+            className="text-lg text-textPrimary hover:text-primary"
             href={`https://anilist.co/staff/${seiyuu.id}`}
           >
             {seiyuu.name.full}
