@@ -1,7 +1,7 @@
 import { getSeiyuuAnimeRoles } from 'api/anilist'
 import { Seiyuu, SeiyuuRole } from 'api/anilist/types'
-import LoadingIndicator from 'components/shared/LoadingIndicator'
-import { AuthContext } from 'context/AuthContext'
+import LoadingIndicator from '../base-ui/LoadingIndicator'
+import { AuthContext } from '../AuthContext'
 import { useContext, useEffect, useState } from 'react'
 import { useInView } from 'react-intersection-observer'
 import SeiyuuCard from './SeiyuuCard'
