@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-module.exports = {
-  reactStrictMode: true,
+const nextConfig = {
+  experimental: {
+    appDir: true,
+  },
 }
+
+module.exports = nextConfig
